tests/integration: move index and setup helpers into a common module
`setup`, `add_table`, `make_index`, and the scan/uuid helpers were
defined in `routing.rs` even though `index_status.rs` already needed
them too and had to reach for them.

Move them into a new common module, shared by all integration test
modules, with no behavior change.
