+++ crates/vector-store/tests/integration/main.rs
@@ -3,6 +3,7 @@
  * SPDX-License-Identifier: LicenseRef-ScyllaDB-Source-Available-1.0
  */
 
+mod common;
 mod db_basic;
 mod fts;
 mod https;
